Adjust logs fetching to avoid missing some
A (seemingly) undocumented feature of CloudWatch's GetLogEvents operation is that, if you don't set a startTime in your command request, you will potentially NOT start reading the log events of your LogStream from the start.

This has been causing an issue for a while where the logs of the last occurrence of weekly runs was not being retrieved, despite the correct log streams being explicitly targeted. More recently, it often caused the app not to show the logs of all runs of the newly introduced LysTrigger lambda - which turned out to be annoying when trying to figure out whether the lambda was running as scheduled or not.